updates, loading tracks, asynchronous!!!!

+++ src/main/java/com/github/samophis/lavaclient/entities/AudioLoadResult.java
@@ -31,5 +31,5 @@ public interface AudioLoadResult {
 	@CheckReturnValue
 	@Nullable
 	@Nonnegative
-	Long selectedTrack();
+	Integer selectedTrack();
 }
